:truck: renames indexProgram to indexClinicalData
