+++ src/indexClinicalData/types.ts
